+++ pocket_ic/__init__.py
@@ -1,11 +1,13 @@
 """
-This package contains 'PocketIC' and 'PocketICServer'.
+`PocketIC` represents an IC instance and is the public interface
+of this package.
 
-'PocketIC' represents an IC instance and is the public interface 
-of this package. 
+Instances are running on a PocketIC server, which is represented
+by `PocketICServer`.
 
-Instances are running on a PocketIC server, which is represented 
-by 'PocketICServer'. 
+`SubnetConfig` is used to configure the subnets of a PocketIC instance.
 """
+
 from .pocket_ic import *
 from .pocket_ic_server import *
+from .subnet_config import *