+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 3.1.0 - 2025-04-28
+
+### Added
+- Support for PocketIC server version 9.0.0
+- Load and store PocketIC state from/to a directory 
+
+
+
 ## 3.0.1 - 2025-02-27
 
 ### Added